(refactoring) Move TTestTableDescription to ut_helpers KIKIMR-21006 (#1985)
